// src/modules/onboarding/dto/company-info.dto.ts
export class CompanyInfoDto {
  companyName!: string;
  legalName?: string | null;
  industry?: string | null;
  website?: string | null;
  email?: string | null;
  phoneNumber?: string | null;
  country?: string | null;
  city?: string | null;
  address?: string | null;
  companySize?: string | null;
}